requirements.txt
# SR-MARE Requirements
# Self-Reflective Multi-Agent Research Engine

# Core dependencies
requests>=2.31.0
numpy>=1.24.0
faiss-cpu>=1.7.4
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6

# Optional: Use faiss-gpu instead of faiss-cpu if GPU is available
# faiss-gpu>=1.7.4

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Utilities
python-dateutil>=2.8.2
